Update backend configuration to use git-gateway
Changes the backend from 'github' to 'git-gateway' in the configuration file.

This adjustment improves compatibility with the current deployment setup, but relies on deprecated Netlify Identity service. Note that for github backend, it may be necessary to include the desired user name in the repo field: user/repo, and to define base_url pointing to the Github OAuth API.