Run animations when in view and improve lottie logic when loop is false
Attributes loop and autoplay are set by default, but can be set to false in Front Matter. When autoplay is set, the players play and pause the animation as it comes into view using an IntersectionObserver.

In addition, if loop is false, the animation is automatically rewinded to its start when paused.
This allows effects that are triggered by scrolling in the scroll-telling web site style.

As examples, the atom animation now starts and stop but there is no perceptible change, and the NATIXAR animation, which doesn't loop, now jumps to life at every pass.