Adds Lottie animations and updates section styling
Implements Lottie animations across multiple language sections, enhancing visual appeal and interactivity.

Updates padding and background styles for the hero block to improve layout consistency.

Removes unused testimonial blocks for cleaner code.

Ensures lottie animation source paths start with a "/"  to tell hugo that the media is not relative to the language folder, but to the root. Without the leading "/" , the generated code attempts to load from, e.g.  /fr/media/atom.lottie . This can be useful if the animations contain localized text.